refactor(sync): centralize timeblock sync lifecycle coordinator
