+++ src/components/TimeBlockWidget.tsx
@@ -29,16 +29,12 @@ import {
 import { useToast } from '@/components/ui/toast-hook';
 import { getTimeBlockService, TimerMode, TimerConfig } from '@/lib/services';
 import type { ActiveBlockData } from '@/lib/types/event';
-import { buildSyncErrorLog } from '@/lib/storage/sync-error';
 import {
   DEFAULT_TIMER_END_SOUND_PRESET_ID,
   getTimerEndSoundPresetById,
   TIMER_END_SOUND_PRESETS,
   type TimerEndSoundPresetId,
 } from '@/lib/media/timer-end-sounds';
-import { useSyncStore } from '@/ui/stores/sync-store';
-import { buildRemoteDbUrl } from '@/lib/sync/remote-db-url';
-import { resolveSyncServerUrl, SYNC_SERVER_URL_CHANGED_EVENT } from '@/config/port-env';
 
 interface TimeBlockWidgetProps {
   /** 是否展开高级选项 */
@@ -119,11 +115,6 @@ export const TimeBlockWidget = forwardRef<TimeBlockWidgetHandle, TimeBlockWidget
 
   // Service
   const timeBlockService = getTimeBlockService();
-  const isLoggedIn = useSyncStore((state) => state.isLoggedIn);
-  const currentUser = useSyncStore((state) => state.currentUser);
-  const [syncServerUrl, setSyncServerUrl] = useState(() =>
-    resolveSyncServerUrl(import.meta.env as Record<string, string | undefined>)
-  );
 
   const clearSkipFeedbackConfirmInterval = useCallback(() => {
     if (!skipFeedbackConfirmIntervalRef.current) {
@@ -273,44 +264,6 @@ export const TimeBlockWidget = forwardRef<TimeBlockWidgetHandle, TimeBlockWidget
       unsubscribe();
     };
   }, [applyActiveBlock, timeBlockService]);
-
-  useEffect(() => {
-    const refreshSyncServerUrl = () => {
-      setSyncServerUrl(resolveSyncServerUrl(import.meta.env as Record<string, string | undefined>));
-    };
-
-    refreshSyncServerUrl();
-    window.addEventListener(SYNC_SERVER_URL_CHANGED_EVENT, refreshSyncServerUrl);
-    return () => {
-      window.removeEventListener(SYNC_SERVER_URL_CHANGED_EVENT, refreshSyncServerUrl);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!isLoggedIn || !currentUser) {
-      void timeBlockService.stopSync();
-      return;
-    }
-
-    let cancelled = false;
-    const remoteUrl = buildRemoteDbUrl(syncServerUrl, currentUser);
-
-    void timeBlockService.startSync(remoteUrl).catch((error) => {
-      if (cancelled) return;
-      const [logMessage, payload] = buildSyncErrorLog('TimeBlockWidget', remoteUrl, error);
-      console.error(logMessage, payload);
-      toast({
-        title: '时间块同步启动失败',
-        description: error instanceof Error ? error.message : String(error),
-        variant: 'destructive',
-      });
-    });
-
-    return () => {
-      cancelled = true;
-      void timeBlockService.stopSync();
-    };
-  }, [currentUser, isLoggedIn, syncServerUrl, timeBlockService, toast]);
 
   // 启动定时器
   const startTimer = useCallback(() => {
